Add caching, pagination, redirects, better errors, & --org option to backup all org repos
Using the option ``--org`` treats the positional argument ``<repo>`` as a Git organiztaion in which all repositories belonging to it will
be backedup, including collaborator forks.

The backup file structure is:
```
:path_from_-o:/:timestamp/:org/:repo/:user/:repo/
```

__NOTE:__ Failed http requests to ``GET`` a repo's collaborator list do not raise, instead a warning is given and that repo is skipped.

The following two aids are used to reduce the total number of Git API requests made (API max 5000/hr):

 * ``?per_page=100``
 * responses are cached
Using the option ``--org`` treats the positional argument ``<repo>`` as a Git organiztaion in which all repositories belonging to it will
be backedup, including collaborator forks.

The backup file structure is:
```
:path_from_-o:/:timestamp/:org/:repo/:user/:repo/
```

__NOTE:__ Failed http requests to ``GET`` a repo's collaborator list do not raise, instead a warning is given and that repo is skipped.

The following two aids are used to reduce the total number of Git API requests made (API max 5000/hr):

 * ``?per_page=100``
 * responses are cached

Signed-off-by: James Noss <jnoss@stsci.edu>